pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "duxx-ai"
version = "0.1.0"
description = "Enterprise Agentic AI SDK — Build, fine-tune, orchestrate, and govern AI agents at scale"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
authors = [
    {name = "Duxx AI Contributors"},
]
keywords = ["ai", "agents", "llm", "fine-tuning", "enterprise", "orchestration", "governance"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "pydantic>=2.0",
    "httpx>=0.25",
    "tiktoken>=0.5",
    "opentelemetry-api>=1.20",
    "opentelemetry-sdk>=1.20",
    "rich>=13.0",
    "click>=8.0",
    "jinja2>=3.1",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
finetune = [
    "torch>=2.0",
    "transformers>=4.36",
    "peft>=0.7",
    "trl>=0.7",
    "datasets>=2.14",
    "bitsandbytes>=0.41",
    "accelerate>=0.25",
]
studio = [
    "fastapi>=0.104",
    "uvicorn>=0.24",
    "websockets>=12.0",
]
tools = [
    "yfinance>=0.2",
    "pdfplumber>=0.10",
]
all = ["duxx-ai[finetune,studio,tools]"]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.1",
    "mypy>=1.7",
]

[project.scripts]
duxx-ai = "duxx_ai.cli.main:cli"

[project.urls]
Homepage = "https://github.com/duxxai/duxx-ai"
Documentation = "https://duxxai.com/docs"
Repository = "https://github.com/duxxai/duxx-ai"

[tool.setuptools.packages.find]
include = ["duxx_ai*"]
exclude = ["duxx_ai.cloud*", "duxx_ai.studio.frontend*"]

[tool.setuptools.exclude-package-data]
duxx_ai = [
    "TIER_SPLIT.py",
    "credentials.py",
    "finetune/tinker.py",
    "finetune/storage.py",
    "studio/website.py",
    "studio/frontend/*",
    "cloud/*",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
